+++ src/appleseed/renderer/modeling/material/disneymaterial.cpp
@@ -937,7 +937,7 @@ const DisneyMaterialLayer& DisneyMaterial::get_layer(
 {
     assert(index < get_layer_count());
 
-    if (thread_index == ~0)
+    if (thread_index == size_t(~0))
         return impl->m_layers[index];
 
     assert(thread_index < Impl::MaxThreadCount);
